Removed old docs components
