+++ components/blog/heading.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import GithubSlugger from 'github-slugger';
 
 const PermalinkIcon = () => (
