+++ components/blog/post-components.js
@@ -1,16 +1,80 @@
-import { components as docsComponents } from '../docs/documentation';
+import { H1, H2, H3, H4, H5 } from './text/headings';
+import { Blockquote } from './text/quotes';
+import { Code, InlineCode } from './text/code';
+import { GenericLink } from './text/link';
+import Heading from './heading';
 
-const H1 = ({ children }) => (
-  <h1 className="fw6">
-    {children}
-
+const PostH2 = ({ children, id }) => (
+  <div>
+    <Heading lean id={id}>
+      <H2>{children}</H2>
+    </Heading>
     <style jsx>{`
-      text-align: center;
-      margin-top: 0;
-      font-size: 2rem;
+      div {
+        margin: 40px 0 0 0;
+      }
     `}</style>
-  </h1>
+  </div>
 );
+
+const PostH3 = ({ children, id }) => (
+  <div>
+    <Heading lean id={id}>
+      <H3>{children}</H3>
+    </Heading>
+    <style jsx>{`
+      div {
+        margin: 2rem 0 0 0;
+      }
+    `}</style>
+  </div>
+);
+
+const PostH4 = ({ children, id }) => (
+  <div>
+    <Heading lean id={id}>
+      <H4>{children}</H4>
+    </Heading>
+  </div>
+);
+
+const Details = ({ children }) => {
+  return (
+    <details>
+      {children}
+      <style jsx>
+        {`
+          details {
+            margin: 1rem 0;
+            padding: 0 0.5rem;
+            background: #f9f9f9;
+          }
+          details[open] {
+            overflow: hidden;
+          }
+        `}
+      </style>
+    </details>
+  );
+};
+
+const Summary = ({ children }) => {
+  return (
+    <summary>
+      {children}
+      <style jsx>{`
+        summary {
+          cursor: pointer;
+          outline: none;
+          font-weight: 500;
+        }
+        summary:hover {
+          opacity: 0.8;
+        }
+      `}</style>
+    </summary>
+  );
+};
 
 const Ul = ({ children }) => (
   <ul>
@@ -40,33 +104,6 @@ const Li = ({ children }) => (
   </li>
 );
 
-const Code = ({ children, syntax }) => {
-  return (
-    <pre className={syntax ? ` ${syntax}` : ''}>
-      <code>{children}</code>
-      <style jsx>
-        {`
-          pre {
-            background: #1d1f21;
-            color: #f8f8f2;
-            white-space: pre;
-            overflow: auto;
-            padding: 1.5rem;
-            margin: 40px 0;
-            border-radius: 3px;
-            -webkit-overflow-scrolling: touch;
-          }
-
-          code {
-            font-size: 14px;
-            line-height: 20px;
-          }
-        `}
-      </style>
-    </pre>
-  );
-};
-
 const P = ({ children }) => {
   return (
     <p>
@@ -95,11 +132,19 @@ const Hr = () => (
 );
 
 export const components = {
-  ...docsComponents,
   h1: H1,
+  h2: PostH2,
+  h3: PostH3,
+  h4: PostH4,
+  h5: H5,
+  blockquote: Blockquote,
+  code: Code,
+  inlineCode: InlineCode,
+  a: GenericLink,
+  details: Details,
+  summary: Summary,
   li: Li,
   ul: Ul,
-  code: Code,
   p: P,
   hr: Hr
 };
